feat(executor): salvage spawn + safety gate from retired LangGraph dispatcher (Closes #741)
Create agents/executor.py with fire-and-forget spawn() primitive and
env sanitization (billing-trap). Preserve agents/safety.py as-is.
Delete agents/dispatcher.py (LangGraph graph wrapper) and agents/main.py
(demo). Forward-port billing-trap and unit tests to new executor module.
Remove obsoleted dispatcher-specific tests.
