+++ agents/__init__.py
@@ -8,9 +8,9 @@ See docs/agents/ for setup and operational notes.
 
 __all__ = [
     "config",
-    "dispatcher",
     "escalation",
     "event_monitor",
+    "executor",
     "github_client",
     "ollama_client",
     "safety",
