+++ scripts/observability/morning_check.py
@@ -20,7 +20,7 @@ from collections import defaultdict
 from datetime import datetime, timedelta, timezone
 from typing import Any
 
-from agents.dispatcher import _hash_scope_files, _now_iso
+from agents.executor import _hash_scope_files, _now_iso
 from agents.supabase_client import get_client
 
 # Alarm categories for stable idempotency key generation
